Filter bug fix
